Added: Code for Cron Job and Scheduler

+++ src/main/java/com/bridgelabz/fundoo_notes/service/NoteService.java
@@ -1,6 +1,9 @@
 package com.bridgelabz.fundoo_notes.service;
 
+import java.time.LocalDate;
+import java.time.ZoneId;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
@@ -11,6 +14,7 @@ import org.springframework.stereotype.Service;
 import com.bridgelabz.fundoo_notes.dto.NoteDto;
 import com.bridgelabz.fundoo_notes.entity.Note;
 import com.bridgelabz.fundoo_notes.entity.User;
+import com.bridgelabz.fundoo_notes.exception.NoteException;
 import com.bridgelabz.fundoo_notes.exception.UserException;
 import com.bridgelabz.fundoo_notes.repository.NotesRepository;
 import com.bridgelabz.fundoo_notes.repository.UserRepository;
@@ -36,17 +40,16 @@ public class NoteService implements INoteService {
 
 	@Autowired
 	JwtToken jwtToken;
-	
+
 	@Autowired
 	UserRepository userRepo;
-	
-    ApiResponse userNotFound = new ApiResponse("User Not Found", 601, null);
 
+	ApiResponse userNotFound = new ApiResponse("User Not Found", 601, null);
 
 	@Override
 	public ApiResponse postNote(String token, NoteDto noteDto) {
 		String email = jwtToken.decodeToken(token);
-		User user = userRepo.findByEmail(email).orElseThrow(()-> new UserException());
+		User user = userRepo.findByEmail(email).orElseThrow(() -> new NoteException());
 		if (user.getNotes() == null) {
 			user.setNotes(new ArrayList<>());
 		}
@@ -54,9 +57,9 @@ public class NoteService implements INoteService {
 		user.getNotes().add(note);
 
 		note.setUser(user);
-		
+
 		noteRepo.save(note);
-		
+
 		apiResponse = new ApiResponse(environment.getProperty("note.addedSuccessfully"), 1, null);
 
 		return apiResponse;
@@ -70,34 +73,79 @@ public class NoteService implements INoteService {
 
 	@Override
 	public ApiResponse updateNote(Integer id, NoteDto noteDto) {
-		Note note = noteRepo.findById(id).orElseThrow(()-> new UserException());
-        if (note != null) {
-            if (noteDto.getTitle() != null) {
-                note.setTitle(noteDto.getTitle());
-            }
-            if (noteDto.getNoteBody() != null) {
-                note.setNoteBody(noteDto.getNoteBody());
-            }
-            noteRepo.save(note);
-            
-        	apiResponse = new ApiResponse(environment.getProperty("note.updateSuccessfully"), 1, null);
-        	
-        	return apiResponse;
-        	
-        } else {
-        	apiResponse = new ApiResponse(environment.getProperty("user.invalidDetails"), 2, null);
-    	
-        	return apiResponse;
-        }
+		Note note = noteRepo.findById(id).orElseThrow(() -> new NoteException());
+		if (note != null) {
+			if (noteDto.getTitle() != null) {
+				note.setTitle(noteDto.getTitle());
+			}
+			if (noteDto.getNoteBody() != null) {
+				note.setNoteBody(noteDto.getNoteBody());
+			}
+			noteRepo.save(note);
+
+			apiResponse = new ApiResponse(environment.getProperty("note.updateSuccessfully"), 1, null);
+
+			return apiResponse;
+
+		} else {
+			apiResponse = new ApiResponse(environment.getProperty("user.invalidDetails"), 2, null);
+
+			return apiResponse;
+		}
 	}
 
 	@Override
-	public ApiResponse deleteNote(Integer id) {
-		Note note = noteRepo.findById(id).orElseThrow(()-> new UserException());
-		if (note != null) {
+	public ApiResponse deleteNote(Integer id, String token) {
+		Note note = noteRepo.findById(id).orElseThrow(() -> new NoteException());
+		String email = jwtToken.decodeToken(token);
+
+		User user = userRepo.findByEmail(email).orElseThrow(() -> new UserException());
+		if (note != null && user != null) {
 			noteRepo.deleteById(id);
 			apiResponse = new ApiResponse("Note Successfully Deleted", 1, note);
 		}
 		return apiResponse;
+	}
+
+	@Override
+	public ApiResponse trashNote(Integer id, String token) {
+		String email = jwtToken.decodeToken(token);
+		Note note = noteRepo.findById(id).orElseThrow(() -> new NoteException());
+		User user = userRepo.findByEmail(email).orElseThrow(() -> new UserException());
+		if (note != null && user != null) {
+			note.setInTrash(true);
+			note.setTrashedDate(new Date());
+		}
+		noteRepo.save(note);
+		apiResponse = new ApiResponse("Note successfully moved to trash", 1, null);
+
+		return apiResponse;
+	}
+
+	@Override
+	public ApiResponse archieveNote(Integer id, String token) {
+		String email = jwtToken.decodeToken(token);
+		Note note = noteRepo.findById(id).orElseThrow(() -> new NoteException());
+		User user = userRepo.findByEmail(email).orElseThrow(() -> new UserException());
+		if (note != null && user != null) {
+			note.setInArchieve(true);
+		}
+		noteRepo.save(note);
+		apiResponse = new ApiResponse("Note successfully moved to Archieves", 1, null);
+
+		return apiResponse;
+	}
+
+	@Override
+	public void deleteTrashedNote() {
+		java.util.List<Note> noteList = noteRepo.findAll();
+		noteList.stream().filter(note -> note.isInTrash()).forEach(note -> {
+			LocalDate date = note.getTrashedDate().toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
+			LocalDate currentDate = new Date().toInstant().atZone(ZoneId.systemDefault()).toLocalDate();
+			LocalDate date2 = date.plusDays(7);
+			if (date2.compareTo(currentDate) < 0) {
+				noteRepo.deleteById(note.getId());
+			}
+		});
 	}
 }
